fix ApiKey page
