+++ src/pages/developers/ApiKeys.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 
 import { useCurrentUserInfosVar, addToast } from '~/core/apolloClient'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { Typography, Button, Icon, Avatar } from '~/components/designSystem'
+import { Typography, Button, Avatar, Icon } from '~/components/designSystem'
 import { HEADER_TABLE_HEIGHT, NAV_HEIGHT, theme } from '~/styles'
 
 gql`
@@ -18,58 +18,76 @@ const ApiKeys = () => {
   const { translate } = useInternationalization()
 
   return (
-    <Page>
-      <Typography variant="headline" color="grey700">
-        {translate('text_6227a2e847fcd700e903893f')}
-      </Typography>
-      <Subtitle variant="body" color="grey600">
-        {translate('text_6227a2e847fcd700e9038947')}
-      </Subtitle>
-      <CopyBlockTitle variant="bodyHl" color="grey500">
-        {translate('text_6227a2e847fcd700e903893f')}
-      </CopyBlockTitle>
-      <CopyBlock>
-        <CopyBlockLeft>
-          <CopyIcon variant="connector" size="medium">
+    <>
+      <ApiKey>
+        <Typography variant="headline">{translate('text_6227a2e847fcd700e903893f')}</Typography>
+        <Subtitle>{translate('text_6227a2e847fcd700e9038947')}</Subtitle>
+        <SubtitleSecretKey variant="bodyHl" color="disabled">
+          {translate('text_6227a2e847fcd700e903893f')}
+        </SubtitleSecretKey>
+        <CopyBlock>
+          <Avatar variant="connector" size="medium">
             <Icon color="dark" name="key" />
-          </CopyIcon>
-          <Typography variant="body" color="grey700">
-            {currentOrganization?.apiKey}
-          </Typography>
-        </CopyBlockLeft>
-        <Button
-          variant="quaternary"
-          size="large"
-          startIcon="duplicate"
-          onClick={() => {
-            navigator.clipboard.writeText(currentOrganization?.apiKey || '')
-            addToast({
-              severity: 'info',
-              translateKey: 'text_6227a2e847fcd700e9038952',
-            })
-          }}
-        >
-          {translate('text_6227a2e847fcd700e903894f')}
-        </Button>
-      </CopyBlock>
-    </Page>
+          </Avatar>
+          <SecretKey color="textSecondary">{currentOrganization?.apiKey}</SecretKey>
+          <Button
+            variant="quaternary"
+            size="large"
+            startIcon="duplicate"
+            onClick={() => {
+              navigator.clipboard.writeText(currentOrganization?.apiKey || '')
+              addToast({
+                severity: 'info',
+                translateKey: 'text_6227a2e847fcd700e9038952',
+              })
+            }}
+          >
+            {translate('text_6227a2e847fcd700e903894f')}
+          </Button>
+        </CopyBlock>
+      </ApiKey>
+      <OrganizationId>
+        <Typography variant="headline">{translate('text_636df520279a9e1b3c68cc75')}</Typography>
+        <SubtitleSecretKey variant="bodyHl" color="disabled">
+          {translate('text_636df520279a9e1b3c68cc75')}
+        </SubtitleSecretKey>
+        <CopyBlock>
+          <Avatar variant="connector" size="medium">
+            <Icon color="dark" name="key" />
+          </Avatar>
+          <SecretKey color="textSecondary">{currentOrganization?.id}</SecretKey>
+          <Button
+            variant="quaternary"
+            size="large"
+            startIcon="duplicate"
+            onClick={() => {
+              navigator.clipboard.writeText(currentOrganization?.id || '')
+              addToast({
+                severity: 'info',
+                translateKey: 'text_636df520279a9e1b3c68cc7d',
+              })
+            }}
+          >
+            {translate('text_6227a2e847fcd700e903894f')}
+          </Button>
+        </CopyBlock>
+      </OrganizationId>
+    </>
   )
 }
 
-const Page = styled.div`
-  padding: ${theme.spacing(8)} ${theme.spacing(12)};
+const ApiKey = styled.div`
+  padding: ${theme.spacing(12)};
+`
+
+const OrganizationId = styled.div`
+  padding-right: ${theme.spacing(12)};
+  padding-left: ${theme.spacing(12)};
+  padding-bottom: ${theme.spacing(12)};
 `
 
 const Subtitle = styled(Typography)`
-  margin-bottom: ${theme.spacing(8)};
-`
-
-const CopyBlockTitle = styled(Typography)`
-  display: flex;
-  align-items: center;
-  height: ${HEADER_TABLE_HEIGHT}px;
-  width: 100%;
-  box-shadow: ${theme.shadows[7]};
+  margin-top: ${theme.spacing(2)};
 `
 
 const CopyBlock = styled.div`
@@ -78,17 +96,23 @@ const CopyBlock = styled.div`
   box-shadow: ${theme.shadows[7]};
   display: flex;
   align-items: center;
-  justify-content: space-between;
-  margin-bottom: ${theme.spacing(12)};
+
+  > *:first-child {
+    margin-right: ${theme.spacing(3)};
+  }
 `
 
-const CopyBlockLeft = styled.div`
+const SubtitleSecretKey = styled(Typography)`
+  height: ${HEADER_TABLE_HEIGHT}px;
+  width: 100%;
   display: flex;
   align-items: center;
+  margin-top: ${theme.spacing(8)};
+  box-shadow: ${theme.shadows[7]};
 `
 
-const CopyIcon = styled(Avatar)`
-  margin-right: ${theme.spacing(3)};
+const SecretKey = styled(Typography)`
+  margin-right: auto;
 `
 
 export default ApiKeys
